requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.21.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.2
rsa==4.9
six==1.16.0
smmap==5.0.0
streamlit==1.24.0
streamlit-chat==0.1.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
